set traps after reads are complete
read doesn't catch the ^c if we are trapping ^c which results in the
read still going after we caught ^c and tried to shut down.  since I
have no idea how to kill read, it is just as easy to not set the cleanup
traps until after all the reads are done.  This may leave a stray monitor
mode interface if the user doesn't fully run through the script, but all
of the scripts can handle that already so lesser of two evils.